Fix attachment personalization (#5)
The program allows the use of subsitution patterns (<?>) in the attachment
filename.  However, in practice, the substitution would only be applied once,
and the rest of the recipients would get the first recipient's attachment.

This happened because, when iterating over the recipients, we would
overwrite the `attachment` variable with the personalized value on the
first iteration. Then, for all further iterations, the attachment didn't
look like it needed personalization, because the substitution pattern
was no longer there.